requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.0
celery==5.3.4
redis==5.0.1
langchain==0.0.335
openai==1.2.4
firebase-admin==6.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
chromadb==0.4.18
sendgrid==6.10.0
twilio==8.10.0

